+++ tests/test_utils.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
+import functools
 import gc
 import os
+import warnings
 
 import numpy as np
 import numpy.typing as npt
@@ -17,6 +19,23 @@ def create_test_om_file(filename: str = "test_file.om", shape=(5, 5), dtype: npt
     writer.close(variable)
 
     return filename, test_data
+
+
+def filter_numpy_size_warning(func):
+    """
+    Decorator to filter out numpy size changed warnings.
+
+    for some reason xr.open_dataset triggers a warning:
+    "RuntimeWarning: numpy.ndarray size changed, may indicate binary incompatibility. Expected 16 from C header, got 96 from PyObject"
+    We will just filter it out for now...
+    https://github.com/pydata/xarray/issues/7259
+    """
+    @functools.wraps(func)
+    def wrapper(*args, **kwargs):
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", message="numpy.ndarray size changed", category=RuntimeWarning)
+            return func(*args, **kwargs)
+    return wrapper
 
 
 class FileDescriptorCounter: